# ExplorerFrontend/app/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZondScan Public API",
    "version": "1.1.0",
...
  },
  "servers": [{ "url": "https://zondscan.com" }],
...
  "paths": {
...
    "/api/tx/{query}": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getTransaction",
        "summary": "Transaction detail",
        "description": "Full detail for one confirmed transaction: transfer record, live receipt status and event logs, calldata, token transfers, internal calls, and contract metadata for decoding. This endpoint is deliberately uncached so a newly mined transaction shows its true confirmation count immediately.",
        "parameters": [
          { "name": "query", "in": "path", "required": true, "description": "Transaction hash: 0x followed by 64 hex characters.", "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{64}$" } }
        ],
        "responses": {
          "200": {
            "description": "Transaction detail. Optional fields appear only when the transaction has the corresponding data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "$ref": "#/components/schemas/Transfer" },
                    "latestBlock": { "type": "integer", "description": "Latest indexed block height as a decimal integer." },
                    "receiptStatus": { "type": "string", "description": "EVM receipt status: 0x1 success, 0x0 reverted. Absent when the receipt fetch failed." },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "address": { "type": "string" },
                          "topics": { "type": "array", "items": { "type": "string" } },
                          "data": { "type": "string" },
                          "logIndex": { "type": "string" },
                          "removed": { "type": "boolean" },
                          "contract": { "$ref": "#/components/schemas/ContractMeta" }
                        }
                      }
                    },
                    "input": { "type": "string", "description": "Transaction calldata as a 0x-prefixed hex string. Present for contract calls." },
                    "targetContract": { "$ref": "#/components/schemas/ContractMeta" },
                    "contractCreated": {
                      "type": "object",
                      "properties": {
                        "address": { "type": "string" },
                        "isToken": { "type": "boolean" },
                        "name": { "type": "string" },
                        "symbol": { "type": "string" },
                        "decimals": { "type": "integer" },
                        "tokenStandard": { "type": "string" }
                      }
                    },
                    "tokenTransfers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "contractAddress": { "type": "string" },
                          "from": { "type": "string" },
                          "to": { "type": "string" },
                          "amount": { "type": "string" },
                          "tokenName": { "type": "string" },
                          "tokenSymbol": { "type": "string" },
                          "tokenDecimals": { "type": "integer" },
                          "tokenStandard": { "type": "string" },
                          "tokenID": { "type": "string" },
                          "logIndex": { "type": "string" }
                        }
                      }
                    },
                    "internalTransactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": { "type": "string" },
                          "callType": { "type": "string" },
                          "from": { "type": "string" },
                          "to": { "type": "string" },
                          "input": { "type": "string" },
                          "output": { "type": "string" },
                          "value": { "type": "string" },
                          "gas": { "type": "string" },
                          "gasUsed": { "type": "string" },
                          "traceAddress": { "type": "array", "items": { "type": "integer" } }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "response": {
                    "ID": "665f1c2ab3d4e5f60718293a",
                    "BlockNumber": "0x24420",
                    "BlockTimestamp": "0x6a743280",
                    "From": "QaaaAAaaaAAaAaaAaAAAAaAAAaAaAaAAaAaaAaaaaAAAAAAAAaAAAAaAaAAaaAAaaaaAaAAAAaaAaAAaaaaaaAaAAaaaaAaAaaaaAaaaAAAAaAAAAaAaAaaAAAaAaaaAA",
                    "To": "QBbBbbBBbBbBBBBbBbBbBBbBBBbBBBbBbBBBbbbBBbBBbBbbBBBBbBBbbbbbBbbBBbbBBBbbBBbbbBbBBbBBbbbbBBbBBBbBbBBbbBbbBBbBBBBbbBBbbbbBBBbbBbbBb",
                    "TxHash": "0x9c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d",
                    "Value": "0x2386f26fc10000",
                    "GasUsed": "0x5208",
                    "GasPrice": "0x3b9aca00",
                    "Nonce": "0x4",
                    "Size": "0x2d0"
                  },
                  "latestBlock": 148512,
                  "receiptStatus": "0x1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction hash format.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Transaction not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "InternalError": {
        "description": "Internal server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "Transfer": {
        "type": "object",
        "description": "Stored transfer record for one transaction. BlockNumber, BlockTimestamp, Value, GasUsed, GasPrice, Nonce, and Size are 0x-prefixed hex strings. Signature is the transaction's ML-DSA-87 signature and Pk is the signer's public key.",
        "properties": {
          "ID": { "type": "string", "description": "Internal record id." },
          "BlockNumber": { "type": "string" },
          "BlockTimestamp": { "type": "string" },
          "From": { "type": "string" },
          "To": { "type": "string" },
          "TxHash": { "type": "string" },
          "Value": { "type": "string" },
          "GasUsed": { "type": "string", "description": "Actual gas consumed when the receipt is available, the gas limit otherwise." },
          "GasPrice": { "type": "string" },
          "Nonce": { "type": "string" },
          "Signature": { "type": "string" },
          "Pk": { "type": "string" },
          "Size": { "type": "string" },
          "Input": { "type": "string", "description": "Calldata as a 0x-prefixed hex string. Present for contract calls." }
        }
      },
...
      "ContractMeta": {
        "type": "object",
        "description": "Compact contract metadata attached to logs and call targets. Verified records carry provenanceStatus. The abi field is suppressed when recorded compiler or source-bundle metadata is invalid; legacy-unrecorded ABI remains available with an explicit warning state.",
        "properties": {
          "name": { "type": "string" },
          "symbol": { "type": "string" },
          "tokenStandard": { "type": "string" },
          "verified": { "type": "boolean" },
          "contractName": { "type": "string" },
          "abi": { "type": "string" },
          "provenanceStatus": {
            "type": "string",
            "enum": ["digest-backed", "legacy-unrecorded", "invalid-recorded"],
            "description": "Trust classification over compact recorded compiler provenance and versioned source-bundle digest metadata. Compact reads validate digest syntax and reserve exact source recomputation for full contract reads."
          }
        }
      },
...
    }
  }
}