# ExplorerFrontend/app/types/transaction.ts
import type { SVGProps } from 'react';

/**
 * Transaction type enumeration
 */
export enum TransactionType {
  Coinbase = 0,
  Attest = 1,
  Transfer = 2,
  Stake = 3
}

/**
 * Base transaction fields that are always present
 */
interface BaseTransaction {
  InOut: number;
  TxType: TransactionType;
  TxHash: string;
  TimeStamp: number;
  Amount: number | string;
  PaidFees?: number;
  gasUsed?: string;
  gasPrice?: string;
  gasUsedStr?: string;
  gasPriceStr?: string;
}

/**
 * Additional fields that might be present on a transaction
 */
interface OptionalTransactionFields {
  ID?: string;
  id?: string;
  Address?: string;
  TransactionAddress?: string;
  From?: string;
  To?: string;
  Type?: string;
  [key: string]: string | number | undefined;
}

/**
 * Full transaction type combining base and optional fields
 */
export type Transaction = BaseTransaction & OptionalTransactionFields;

/**
 * Base internal transaction fields
 */
interface BaseInternalTransaction {
  Type: number;
  CallType: string;
  Hash: string;
  From: string;
  To: string;
  Input: string;
  Output: number;
  Value: number;
  Gas: number;
  GasUsed: number;
  AddressFunctionIdentifier: string;
  AmountFunctionIdentifier: number;
  BlockTimestamp: number;
}

/**
 * Additional fields for internal transactions
 */
interface OptionalInternalTransactionFields {
  TraceAddress: Array<number | string>;
  [key: string]: string | number | Array<number | string> | undefined;
}

/**
 * Full internal transaction type
 */
export type InternalTransaction = BaseInternalTransaction & OptionalInternalTransactionFields;

/**
 * Token standard tag persisted on contract + transfer rows. Empty for
 * legacy un-classified rows; new rows are always one of the three.
 */
export type TokenStandard = 'ERC-20' | 'ERC-721' | 'ERC-1155';

export type ContractMetadataProvenanceStatus =
  | 'digest-backed'
  | 'legacy-unrecorded'
  | 'invalid-recorded';

/**
 * Token transfer information for a transaction
 */
export interface TokenTransferInfo {
  contractAddress: string;
  from: string;
  to: string;
  amount: string;
  tokenName: string;
  tokenSymbol: string;
  tokenDecimals: number;
  /** ERC-20 | ERC-721 | ERC-1155, drives the row's badge in transaction-view. */
  tokenStandard?: TokenStandard | string;
  /** uint256 decimal string. Populated for ERC-721 + ERC-1155 transfers. */
  tokenID?: string;
  /** Hex log index, used as a stable React key when a tx emits several events. */
  logIndex?: string;
}

/**
 * Detailed transaction information for transaction detail pages
 */
export interface TransactionDetails {
  hash: string;
  blockNumber?: string | number;
  from: string;
  to: string;
  value: string;
  timestamp: number;
  status?: string;
  gasUsed?: string;
  gasPrice?: string;
  nonce?: number;
  latestBlock?: number;
  PaidFees?: number;
  contractCreated?: {
    address: string;
    isToken: boolean;
    name: string;
    symbol: string;
    decimals: number;
    tokenStandard?: TokenStandard | string;
  };
  tokenTransfers?: TokenTransferInfo[];
  /** Raw calldata for the tx, "0x..." prefixed. Empty / "0x" for plain transfers. */
  input?: string;
  /** Receipt logs in emission order. Populated best-effort; absent when the RPC fetch fails. */
  logs?: TxLog[];
  /** Verified contract metadata for the tx's target, used by the Input Data card to decode method calls. */
  targetContract?: ContractMeta;
  /** Internal calls captured under this tx by the syncer (CALL / DELEGATECALL / STATICCALL sub-frames). */
  internalTransactions?: InternalTx[];
  /** Receipt-level status from the live RPC. "0x1" = success, "0x0" = reverted, undefined when RPC fetch failed. */
  receiptStatus?: string;
}

/**
 * One internal-call row attached to a tx. Mirrors the backend's
 * models.InternalTx, which mirrors how the syncer writes the
 * `internalTransactionByAddress` collection.
 *
 * Most txs have an empty array; only contract calls with sub-frames
 * (CALL / DELEGATECALL / STATICCALL) get entries here.
 */
export interface InternalTx {
  type: string;
  callType: string;
  from: string;
  to: string;
  input: string;
  output: string;
  value: number;
  gas: string;
  gasUsed: string;
  traceAddress: number[];
}

/**
 * Optional contract metadata attached to a log or as the tx's target,
 * surfaced by the backend on /tx/:hash when contractCode has a row for
 * the address. `abi` is only populated when `verified === true` since
 * an unverified ABI is unreliable, the frontend uses it to decode
 * unknown event signatures + method selectors.
 */
export interface ContractMeta {
  name?: string;
  symbol?: string;
  tokenStandard?: TokenStandard | string;
  verified?: boolean;
  contractName?: string;
  abi?: string;
  /** Trust state computed by the backend's compact recorded-metadata gate. */
  provenanceStatus?: ContractMetadataProvenanceStatus | string;
}

/**
 * One receipt log entry on a confirmed tx. Mirrors the subset returned by
 * the backend's /tx/:hash route, which itself mirrors qrl_getTransactionReceipt.logs[].
 *
 * `contract` is attached server-side from contractCode for the log's
 * emitting address; the event-log decoder uses it as the ABI fallback
 * when topic[0] doesn't match a well-known signature.
 */
export interface TxLog {
  address: string;
  topics: string[];
  data: string;
  logIndex: string;
  removed?: boolean;
  contract?: ContractMeta;
}

/**
 * Pending transaction from the mempool
 */
export interface PendingTransaction {
  accessList: unknown[];
  blockHash: null;
  blockNumber?: string;
  chainId: string;
  from: string;
  gas: string;
  gasPrice: string;
  hash: string;
  input: string;
  maxFeePerGas?: string;
  maxPriorityFeePerGas?: string;
  nonce: string;
  publicKey: string;
  to?: string;
  transactionIndex: null;
  type: string;
  value: string;
  status: 'pending' | 'mined' | 'dropped';
  lastSeen: number;
  createdAt: number;
}

/**
 * Calculate the number of confirmations for a transaction. Clamped to >=0:
 * if the backend's latestBlock is briefly stale relative to the tx's own
 * block, we'd otherwise return a negative count and badge the (mined) tx
 * as "Pending" with "-N Confirmations", confusing UX.
 */
export function getConfirmations(txBlockNumber?: string | number, latestBlock?: number): number | null {
  if (!txBlockNumber || !latestBlock) return null;
  const blockNum = typeof txBlockNumber === 'string' ? parseInt(txBlockNumber) : txBlockNumber;
  return Math.max(0, latestBlock - blockNum + 1);
}

/**
 * Get transaction status based on confirmations. This helper is called from
 * the mined-tx view (the /pending/tx redirect filters pending-status rows
 * upstream), so any non-null confirmations means the tx is in a block →
 * Confirmed. Only the missing-block-data case (null) falls back to Pending.
 */
export function getTransactionStatus(confirmations?: number | null): {
  text: string;
  color: string;
} {
  if (confirmations === null || confirmations === undefined) {
    return { text: 'Pending', color: 'bg-yellow-500' };
  }
  return { text: 'Confirmed', color: 'bg-green-500' };
}

/**
 * SVG icon props type
 */
export type SVGIconProps = SVGProps<SVGSVGElement>;